services/portal/requirements.txt
fastapi
uvicorn
psycopg2-binary
boto3